Breaking out semver util function
